// src/shared/api/models/getApiBundleOffers200ItemsItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiBundleOffers200ItemsItemApplicableService } from './getApiBundleOffers200ItemsItemApplicableService';
import type { GetApiBundleOffers200ItemsItemBundleOfferType } from './getApiBundleOffers200ItemsItemBundleOfferType';

export type GetApiBundleOffers200ItemsItem = {
    id: string;
    seqId: number;
    outletId: number;
    name: string;
    /** @nullable */
    description: string | null;
    /** Services on which bundle offer can be applied */
    applicableService: GetApiBundleOffers200ItemsItemApplicableService;
    bundleOfferType: GetApiBundleOffers200ItemsItemBundleOfferType;
    sellOnline: boolean;
    price: number;
    expiryMonths: number;
    createdAt: string;
    updatedAt: string;
    /** @nullable */
    deletedAt: string | null;
};